+++ coders/mat.c
@@ -746,8 +746,9 @@ MATLAB_KO: ThrowReaderException(CorruptImageError,"ImproperImageHeader");
     }
 
     CellType = ReadBlobXXXLong(image2);    /* Additional object type */
-    if (logging) (void)LogMagickEvent(CoderEvent,GetMagickModule(),
-          "MATLAB_HDR.CellType: %ld",CellType);
+    if (logging)
+      (void) LogMagickEvent(CoderEvent,GetMagickModule(),
+        "MATLAB_HDR.CellType: %ld",(long) CellType);
   
     (void) ReadBlob(image2, 4, (unsigned char *) &size);     /* data size */
