+++ coders/mpeg.c
@@ -540,11 +540,12 @@ static MagickBooleanType WriteMPEGImage(const ImageInfo *image_info,
             *frame;
 
           (void) FormatMagickString(p->filename,MaxTextExtent,"%s%lu.%s",
-            basename,p->scene,WriteMPEGIntermediateFormat);
+            basename,(unsigned long) p->scene,WriteMPEGIntermediateFormat);
           (void) FormatMagickString(filename,MaxTextExtent,"%s%lu.%s",
-            basename,p->scene,WriteMPEGIntermediateFormat);
+            basename,(unsigned long) p->scene,WriteMPEGIntermediateFormat);
           (void) FormatMagickString(previous_image,MaxTextExtent,
-            "%s%lu.%s",basename,p->scene,WriteMPEGIntermediateFormat);
+            "%s%lu.%s",basename,(unsigned long) p->scene,
+            WriteMPEGIntermediateFormat);
           frame=CloneImage(p,0,0,MagickTrue,&p->exception);
           if (frame == (Image *) NULL)
             break;
@@ -560,7 +561,7 @@ static MagickBooleanType WriteMPEGImage(const ImageInfo *image_info,
         default:
         {
           (void) FormatMagickString(filename,MaxTextExtent,"%s%lu.%s",
-            basename,p->scene,WriteMPEGIntermediateFormat);
+            basename,(unsigned long) p->scene,WriteMPEGIntermediateFormat);
           if (length > 0)
             status=BlobToFile(filename,blob,length,&image->exception);
           break;
@@ -570,12 +571,12 @@ static MagickBooleanType WriteMPEGImage(const ImageInfo *image_info,
         {
           if (status != MagickFalse)
             (void) LogMagickEvent(CoderEvent,GetMagickModule(),
-              "%lu. Wrote %s file for scene %lu:",i,WriteMPEGIntermediateFormat,
-              p->scene);
+              "%lu. Wrote %s file for scene %lu:",(unsigned long) i,
+              WriteMPEGIntermediateFormat,(unsigned long) p->scene);
           else
             (void) LogMagickEvent(CoderEvent,GetMagickModule(),
-              "%lu. Failed to write %s file for scene %lu:",i,
-              WriteMPEGIntermediateFormat,p->scene);
+              "%lu. Failed to write %s file for scene %lu:",(unsigned long) i,
+              WriteMPEGIntermediateFormat,(unsigned long) p->scene);
           (void) LogMagickEvent(CoderEvent,GetMagickModule(),"%s",
             filename);
         }
@@ -612,7 +613,7 @@ static MagickBooleanType WriteMPEGImage(const ImageInfo *image_info,
     for (i=0; i < (ssize_t) MagickMax((1.0*delay+1.0)/3.0,1.0); i++)
     {
       (void) FormatMagickString(p->filename,MaxTextExtent,"%s%lu.%s",
-        basename,count++,WriteMPEGIntermediateFormat);
+        basename,(unsigned long) count++,WriteMPEGIntermediateFormat);
       (void) RelinquishUniqueFileResource(p->filename);
     }
     (void) CopyMagickString(p->filename,image_info->filename,MaxTextExtent);
